check source path exists when copying to kubemaster (#426)

+++ lib/modules/k2s/k2s.node.module/linuxnode/vm/vm.module.psm1
@@ -192,7 +192,7 @@ function Copy-ToControlPlaneViaSSHKey($Source, $Target,
     Write-Log "copy: $Source to: $Target IgnoreErrors: $IgnoreErrors"
     $error.Clear()
     $leaf = Split-Path $Source -leaf
-    if ((Get-Item $Source) -is [System.IO.DirectoryInfo] -and $leaf -ne "*") {
+    if ($(Test-Path $Source) -and (Get-Item $Source) -is [System.IO.DirectoryInfo] -and $leaf -ne "*") {
         # is directory
         Invoke-CmdOnControlPlaneViaSSHKey "sudo rm -rf /tmp/copy.tar"
         $leaf = Split-Path $Source -Leaf
